requirements.txt
Flask_WTF==1.2.1
Werkzeug==3.0.2
Flask==3.0.3
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.29
Jinja2==3.1.3
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.8.4